Refactor difficulty system to use enum class
Replaced legacy integer difficulty macros with strongly-typed uqm::Difficulty enum class. Updated all gameplay logic, options, menu handling, and save/load routines to use new inline functions and enum values. Refactored save/load code for improved type safety and reduced duplication. Enhanced logging and formatting for enums. Added platform abstraction files and environment variable retrieval. Cleaned up old macros and improved code consistency throughout.

+++ src/uqm/sis.cpp
@@ -232,7 +232,7 @@ void DrawSaveInfo(SIS_STATE SisState)
 
 		utf8StringCopy(TempDiff, sizeof(TempDiff),
 					   GAME_STRING(MAINMENU_STRING_BASE + 56
-								   + SisState.Difficulty));
+								   + static_cast<int>(SisState.Difficulty)));
 
 		fmt::format_to_sz_n(buf, sizeof buf, "{} {}{}{}",
 							GAME_STRING(MAINMENU_STRING_BASE + 55), // Difficulty:
